Add DEPENDS.txt to track dependencies for TeX Live
This file is specifically supported by TeX Live to track the
dependencies of a package, and the TL dependency information is used by
other tools. Declaring the dependencies here makes that information
available to tools - in particular, the dependency on `ifmtarg`, which
is not included by default as a package in a LaTeX distribution.

The packages `calc` and `ifthen` are both included by default in every
LaTeX distribution, but there's not much guidance online about if you
need to include such packages or not. Since they are `Require`d, I've
included them.